fix 3box comment issue where only one comment was working
