+++ app/src/components/common/index.ts
@@ -27,7 +27,6 @@ export { SubsectionTitleWrapper } from './text/subsection_title_wrapper'
 export { TD, TH, THead, TR, Table } from './tables/table'
 export { Textfield } from './form/textfield'
 export { TextfieldCustomSymbol as TextfieldCustomPlaceholder } from './form/textfield_custom_symbol'
-// export { ThreeBoxComments } from './three_box_comments'
 export { TitleValue } from './text/title_value'
 export { Tooltip } from './tooltip'
 export { Well } from './text/well'
